Cleanup implementation
